@@ -1,6 +1,6 @@
-module github.com/rwestlund/quickbooks-go
+module github.com/tommyhedley/quickbooks-go
 
-go 1.20
+go 1.23.0
 
 require (
 	github.com/stretchr/testify v1.9.0
